# doc/api.yml
swagger: '2.0'
info:
  title: Guido
  description: PLD-SMART project API
  version: "0.0.1"
host: localhost
schemes:
- http
- https
basePath: /v1
produces:
- application/json
paths:
  /comments/{id}:
    get:
      summary: Route Comments
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: A
      responses:
        200:
          description: A
  /events/{id}:
    get:
      summary: Ongoing Events
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: A
      responses:
        200:
          description: A
  /landmarks/{id}:
    get:
      summary: Landmark Data
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: UUID of the desired landmark.
      responses:
        200:
          description: Information about a specific landmark.
          schema:
            $ref: '#/definitions/Landmark'
  /regions/{id}:
    get:
      summary: Regional Data
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: A
      responses:
        200:
          description: A
  /routes:
    post:
      summary: Create a new route
      parameters:
      - name: body
        in: body
        required: true
        description: The name of the route
        schema:
          $ref: '#/definitions/CreateRouteOptions'
      responses:
        200:
          description: Route was successfully updated.
          schema:
            type: object
            description: Response string.
            properties:
              status:
                type: string
                description: A status string.
  /routes/{id}:
    get:
      summary: Route Information
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: UUID of the desired route.
      responses:
        200:
          description: Information about a specific route.
          schema:
            $ref: '#/definitions/Route'
    put:
      summary: Update Route Information
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: UUID of the desired route.
      responses:
        200:
          description: Route was successfully updated.
          schema:
            type: object
            description: Response string.
            properties:
              status:
                type: string
                description: A status string.
  /waypoints/{id}:
    get:
      summary: Waypoint Information
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: UUID of the desired waypoint.
      responses:
        200:
          description: Information about a specific waypoint.
          schema:
            $ref: '#/definitions/Waypoint'
  /users/{id}:
    get:
      summary: User Profile
      description: Summary of the user's static details.
      parameters:
      - name: id
        in: path
        required: true
        type: string
        format: uuid
        description: UUID of the desired user.
      responses:
        200:
          description: User summary.
definitions:
  Landmark:
    type: object
    description: Geographical point of interest.
    properties:
      title:
        type: string
        description: Title of the landmark.
      latitude:
        type: number
        description:  Geographical latitude of the landmark
      longitude:
        type: number
        format: double
        description: Geographical longitude of the landmark.
  CreateRouteOptions:
    type: object
    description: The parameters to create a new route
    properties:
      title:
        type: string
        description: The title of the new route
      description:
        type: string
        description: The description of the new route
      duration:
        type: integer
        format: int32
        description: The duration of the route in minutes
  Route:
    type: object
    description: Complete description of a route.
    properties:
      id:
        type: string
        format: uuid
        description: Unique identifier representing a specific route.
      title:
        type: string
        description: Title of the route.
      description:
        type: string
        description: Description of the route.
      author:
        $ref: '#/definitions/User'
      creation_date:
        type: string
        format: 'date-time'
        description: Date and time of creation of this route.
      modification_date:
        type: string
        format: 'date-time'
        description: Date and time of the last modification to this route.
      duration:
        type: integer
        description: Estimated duration of the route (in minutes).
      likes:
        type: integer
        description: Number of times this route was liked.
      favorites:
        type: integer
        description: Number of times this route was favorited.
      tags:
        type: array
        description: Set of user tags describing the route.
        items:
          type: string
      waypoints:
        type: array
        description: Sequence of waypoints which make up this route.
        items:
          $ref: '#/definitions/Waypoint'
  RouteInfo:
    type: object
    description: Partial description of a route, which omits listing waypoints.
    properties:
      id:
        type: string
        format: uuid
        description: Unique identifier representing a specific route.
      title:
        type: string
        description: Title of the route.
      description:
        type: string
        description: Description of the route.
      imageUrl:
        type: string
        description: URL for the image of the route
      author:
        $ref: '#/definitions/User'
      creation_date:
        type: string
        format: 'date-time'
        description: Date and time of creation of this route.
      modification_date:
        type: string
        format: 'date-time'
        description: Date and time of the last modification to this route.
      duration:
        type: integer
        description: Estimated duration of the route (in minutes).
      likes:
        type: integer
        description: Number of times this route was liked.
      favorites:
        type: integer
        description: Number of times this route was favorited.
      tags:
        type: array
        description: Set of user tags describing the route.
        items:
          type: string
      comments:
        type: array
        description: A list of comments
        items:
          type: string
  User:
    type: object
    description: User of Guido.
    properties:
      id:
        type: string
        format: uuid
        description: Unique identifier representing a specific user.
      username:
        type: string
        description: Alphanumeric identifier for the user.
      favorites:
        type: array
        items:
          $ref: '#/definitions/RouteInfo'
      likes:
        type: array
        items:
          $ref: '#/definitions/RouteInfo'
  Waypoint:
    type: object
    description: Geographical point on a route.
    properties:
      id:
        type: string
        format: uuid
        description: Unique identifier representing a specific waypoint.
      latitude:
        type: number
        format: double
        description: Geographical latitude of the waypoint.
      longitude:
        type: number
        format: double
        description: Geographical longitude of the waypoint.
      note:
        type: string
        description: Short note (255 characters at most) providing additional information about this waypoint.
      duration:
        type: integer
        description: Suggested amount of time to be spent at this way point (in minutes).